Day 9 part 1.

+++ src/day2/Solution.java
@@ -16,11 +16,11 @@ public class Solution {
 
 	public static final String INPUT = "1,0,0,3,1,1,2,3,1,3,4,3,1,5,0,3,2,1,6,19,1,19,6,23,2,23,6,27,2,6,27,31,2,13,31,35,1,9,35,39,2,10,39,43,1,6,43,47,1,13,47,51,2,6,51,55,2,55,6,59,1,59,5,63,2,9,63,67,1,5,67,71,2,10,71,75,1,6,75,79,1,79,5,83,2,83,10,87,1,9,87,91,1,5,91,95,1,95,6,99,2,10,99,103,1,5,103,107,1,107,6,111,1,5,111,115,2,115,6,119,1,119,6,123,1,123,10,127,1,127,13,131,1,131,2,135,1,135,5,0,99,2,14,0,0";
 
-	public static int[] part2(int[] stack){
+	public static long[] part2(long[] stack){
 
 		outer: for(int i = 0; i < 100; i++){
 			for(int j = 0; j < 100; j++){
-				int[] result = stack.clone();
+				long[] result = stack.clone();
 				result[1] = i;
 				result[2] = j;
 				try {
@@ -50,15 +50,15 @@ public class Solution {
 
 		// Process input to stack
 		String[] sep = input.split(",");
-		int[] ogStack = new int[sep.length];
+		long[] ogStack = new long[sep.length];
 		try {
 			// Parse input to numbers
 			for(int i = 0; i < sep.length; i++){
-				ogStack[i] = Integer.parseInt(sep[i]);
+				ogStack[i] = Long.parseLong(sep[i]);
 			}
 
 			// Perform part 1 (running the input)
-			int[] stack = ogStack.clone();
+			long[] stack = ogStack.clone();
 			stack[1] = PART1_X;
 			stack[2] = PART1_Y;
 			IntcodeComputer ic = new IntcodeComputer(stack);
